Move StatementResource to its own package
